+++ src/Data/Singletons/TypeLits/Internal.hs
@@ -16,7 +16,7 @@
 {-# LANGUAGE PolyKinds, DataKinds, TypeFamilies, FlexibleInstances,
              UndecidableInstances, ScopedTypeVariables, RankNTypes,
              GADTs, FlexibleContexts, TypeOperators, ConstraintKinds,
-             TypeInType, TemplateHaskell #-}
+             TypeInType, TemplateHaskell, StandaloneDeriving #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
 module Data.Singletons.TypeLits.Internal (
@@ -156,3 +156,28 @@ sError sstr = error (T.unpack (fromSing sstr))
 type a :^ b = a ^ b
 infixr 8 :^
 $(genDefunSymbols [''(:^)])
+
+------------------------------------------------------------
+-- TypeLits singleton non-singleton instances
+------------------------------------------------------------
+
+-- Thanks to @cumber on #179
+
+instance Show (SNat n) where
+  showsPrec p n@SNat
+    = showParen (p > atPrec)
+      ( showString "SNat @"
+        . showsPrec (atPrec + 1) (natVal n)
+      )
+    where atPrec = 10
+
+instance Show (SSymbol s) where
+  showsPrec p s@SSym
+    = showParen (p > atPrec)
+      ( showString "SSym @"
+        . showsPrec (atPrec + 1) (symbolVal s)
+      )
+    where atPrec = 10
+
+deriving instance Show (SomeSing Nat)
+deriving instance Show (SomeSing Symbol)